% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chess.openings}
\alias{chess.openings}
\title{Chess Openings in Lichess Opening Tournament Schedule}
\format{
A data table with 131 rows and 3 columns:
\itemize{
  \item{Index, as in original order}
  \item{ECO, ECO (The Encyclopaedia of Chess Openings) codes}
  \item{Opening, Opening names}
}
}
\source{
\url{https://benediktwerner.github.io/lichess-themed-tournament-schedule/}
}
\usage{
data(chess.openings)
}
\value{
data.table
}
\description{
Chess Openings in Lichess Opening Tournament Schedule
}
\details{
A dataset containing the original index, ECO codes and opening names
of 131 Chess Openings in Lichess Opening Tournament Schedule
}
\author{

}
\keyword{datasets}
