% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmisc.data-package.R
\docType{package}
\name{Xmisc.data-package}
\alias{Xmisc.data-package}
\title{Xiaobei's collected datasets for R projects}
\description{
Xiaobei's collected datasets for R projects
}
\details{
Listed are a few highlighted sets
\itemize{
\item \code{\link{chess.openings}}
}

You may find the latest version of \code{Xmisc.data} at \url{https://github.com/xiaobeizhao/Xmisc.data}.
}
\references{
Xiaobei Zhao (2020)
}
\author{
Xiaobei Zhao
}
\keyword{package}
